+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### :zap: Added
+
+- [#57](https://github.com/FantasticFiasco/action-update-license-year/issues/57) Support for configuring pull request
+
 ## [1.2.2] - 2020-07-22
 
 ### :syringe: Fixed